ingestion_service/requirements.txt
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
pika==1.3.2
langchain==0.3.26
langchain-openai==0.2.14
langchain-pinecone==0.2.5
langchain-text-splitters==0.3.8
pinecone
